Added SortableHtml renderer:
SortableHtml requires some magic to be passed in for now. Like, url, any url params you want to keep, the sorted column, and the current sort order.

This doesn't actually perform any sorting. It links up the header in the table in the standard 'click a column header and toggle the sort order'. Currently, I am using the information to pass into the order by clause of the SQL. Next up: tell the Report to sort if not the SQL.

I also added column_data_fields (not crazy about the name) to Report so you can map a column name in the report to the underlying table column. This is useful for join models that just don't seem to come through correctly.